pyproject.toml
[project]
name = "chunkweaver"
version = "0.3.0"
description = "RAG chunker that respects document structure."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "Metawake"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Text Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["chunking", "rag", "text-splitting", "vector-database", "nlp", "retrieval"]

[project.urls]
Homepage = "https://github.com/metawake/chunkweaver"
Repository = "https://github.com/metawake/chunkweaver"
Issues = "https://github.com/metawake/chunkweaver/issues"

[project.optional-dependencies]
cli = []
langchain = ["langchain-text-splitters>=0.2"]
llamaindex = ["llama-index-core>=0.11"]
dev = ["pytest>=7.0", "pytest-cov"]
benchmark = ["sentence-transformers>=3.0,<4", "numpy>=1.24,<2"]
all = ["langchain-text-splitters>=0.2", "llama-index-core>=0.11"]

[project.scripts]
chunkweaver = "chunkweaver.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers"

[tool.hatch.build.targets.wheel]
packages = ["chunkweaver"]

[tool.hatch.build.targets.sdist]
include = [
    "chunkweaver/",
    "tests/",
    "LICENSE",
    "README.md",
    "pyproject.toml",
]
exclude = [
    "benchmark/",
    "benchmarks/",
    "experiments/",
    "assets/",
    "article-medium.md",
    "ROADMAP.md",
    "vision.md",
    "chunkweaver.txt",
    ".gitignore",
]

[tool.ruff]
target-version = "py39"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["chunkweaver"]
